refactor(api-response): declare the response envelope once
The success and failure envelope was declared three times: as
ApiHandlerResponseError on the port, byte identically as ApiResponseError
in the interface layer, and re-derived as ErrorResponseResult and
SuccessResponseResult in the Next adapter. The server built responses from
one definition while clients validated against another, and nothing type
checked that the two agreed.

It now lives in common/application/api-response, which the port can name
without depending on the outer interface layer, and it carries the
validation issues the error schema has always declared.
